# openspec/changes/dcf-ai-feedback-architecture-v1/openapi.yaml
openapi: 3.0.3
info:
  title: DCF AI Feedback Platform API
  version: 0.1.0
  description: |
    Contract baseline for event-triggered feedback lifecycle, AI classification, and supervisor triage.
    Demo mode assumptions: static logins for auth, local-first runtime, SQLite-backed persistence.
    Hugging Face config baseline (adapted from DCF-Intake demo):
    LLM_PROVIDER=huggingface, HF_MODEL=Qwen/Qwen2.5-7B-Instruct,
    HF_API_BASE=https://router.huggingface.co/v1, LLM_REQUEST_TIMEOUT_MS=600000,
    token alias support via HF_API_TOKEN or HUGGINGFACE_API_KEY.
servers:
  - url: https://api.dcf-feedback.local/v1
security:
  - demoAuthHeader: []
...
paths:
  /auth/login:
    post:
      tags: [auth]
      summary: Login with static demo credentials
      operationId: loginDemoUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginIn'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionInfo'
        '401':
          description: Invalid credentials
...
  /forms/templates:
    get:
      tags: [surveys]
      summary: List form templates
      operationId: listFormTemplates
      parameters:
        - in: query
          name: userGroup
          schema:
            $ref: '#/components/schemas/UserGroup'
      responses:
        '200':
          description: Templates list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/FormTemplate'
    post:
      tags: [surveys]
      summary: Create form template
      operationId: createFormTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FormTemplateIn'
      responses:
        '201':
          description: Template created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormTemplate'
  /forms/templates/{templateId}:
    put:
      tags: [surveys]
      summary: Update form template
      operationId: updateFormTemplate
      parameters:
        - in: path
          name: templateId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FormTemplateIn'
      responses:
        '200':
          description: Template updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormTemplate'
  /forms/instances:
    post:
      tags: [surveys]
      summary: Create form instance from trigger context
      operationId: createFormInstance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FormInstanceIn'
      responses:
        '201':
          description: Form instance created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SurveyInstance'
  /forms/instances/{instanceId}/links:
    post:
      tags: [surveys]
      summary: Generate dynamic link token
      operationId: createDynamicLink
      parameters:
        - in: path
          name: instanceId
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Dynamic link generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DynamicLink'
  /forms/instances/{instanceId}/send:
    post:
      tags: [surveys]
      summary: Send survey and return custom link
      operationId: sendSurveyLink
      parameters:
        - in: path
          name: instanceId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Survey link sent and returned for UI display
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendSurveyResponse'
  /forms/tokens/{token}/validate:
    post:
      tags: [surveys]
      summary: Validate dynamic token
      operationId: validateDynamicToken
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Token validation status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenValidation'
  /surveys/{surveyInstanceId}:
    get:
      tags: [surveys]
      summary: Fetch survey form metadata and questions
      operationId: getSurveyInstance
      parameters:
        - $ref: '#/components/parameters/SurveyInstanceId'
      responses:
        '200':
          description: Survey metadata returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SurveyInstance'
        '404':
          description: Survey not found or expired
...
components:
  securitySchemes:
    demoAuthHeader:
      type: apiKey
      in: header
      name: x-demo-user
      description: Static demo login identity key that maps to a predefined role profile.
  parameters:
    SurveyInstanceId:
      name: surveyInstanceId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    UserGroup:
      type: string
      enum: [mandated_reporter, volunteer, attorney, foster_parent]
...
    SurveyQuestion:
      type: object
      required: [id, type, label]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [likert, text, multi_select, dropdown, attestation]
        label:
          type: string
        required:
          type: boolean
        options:
          type: array
          items:
            type: string
    FormTemplateIn:
      type: object
      required: [name, userGroup, questions]
      properties:
        name:
          type: string
        userGroup:
          $ref: '#/components/schemas/UserGroup'
        version:
          type: integer
        questions:
          type: array
          items:
            $ref: '#/components/schemas/SurveyQuestion'
        nextStepsLabel:
          type: string
          default: Next Steps
    FormTemplate:
      allOf:
        - $ref: '#/components/schemas/FormTemplateIn'
        - type: object
          properties:
            templateId:
              type: string
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    FormInstanceIn:
      type: object
      required: [templateId, userGroup]
      properties:
        templateId:
          type: string
        userGroup:
          $ref: '#/components/schemas/UserGroup'
        triggerRef:
          type: string
        expiresAt:
          type: string
          format: date-time
    DynamicLink:
      type: object
      properties:
        instanceId:
          type: string
        token:
          type: string
        url:
          type: string
        expiresAt:
          type: string
          format: date-time
    TokenValidation:
      type: object
      properties:
        valid:
          type: boolean
        reason:
          type: string
    SendSurveyResponse:
      type: object
      properties:
        instanceId:
          type: string
        status:
          type: string
          enum: [sent, queued]
        customLink:
          type: string
        deliveryChannel:
          type: string
          enum: [email, portal, both]
    LoginIn:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    SessionInfo:
      type: object
      properties:
        userId:
          type: string
        username:
          type: string
        role:
          type: string
        officeId:
          type: string
        authenticated:
          type: boolean
    SurveyInstance:
      type: object
      required: [surveyInstanceId, userGroup, expiresAt, questions]
      properties:
        surveyInstanceId:
          type: string
        userGroup:
          $ref: '#/components/schemas/UserGroup'
        triggerRef:
          type: string
        expiresAt:
          type: string
          format: date-time
        aiDisclosureRequired:
          type: boolean
        questions:
          type: array
          items:
            $ref: '#/components/schemas/SurveyQuestion'
...